requirements.txt
torch==2.0.1
torchvision==0.15.2
lightly==1.4.6
PyYAML==6.0
matplotlib==3.7.1
Pillow==9.5.0
scikit-learn==1.2.2
pytorch-lightning==1.7.2
streamlit==1.22.0
mediapipe==0.10.0
opencv-python==4.7.0.72
tensorboard==2.13.0
tensorboardX==2.6
loguru==0.7.0
